requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

asgiref==3.8.1
boto3==1.34.44
botocore==1.34.44
celery==5.3.6
channels==4.0.0
Django==5.1
django-datatables-view==1.20.0
django-ratelimit==4.1.0
django-vectordb==0.2.0
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-api-key==2.3.0
django-constance==3.1.0
dspy-ai==2.4.10
httpx==0.27.0
limits==3.13.0
ninja==1.11.1.1
openai==1.46.1
pgvector==0.3.2
pydantic==2.5.0
python-decouple==3.6
pytz==2023.3.post1
regex==2023.12.25
requests==2.32.3
sentence-transformers==3.0.1
smart-open==7.0.4
stripe==5.0.0
tiktoken==0.6.0
transformers==4.38.1
typing_extensions==4.9.0

